requirements.txt
altair==5.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
boto3==1.34.89
botocore==1.34.89
branca==0.7.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
folium==0.16.0
fonttools==4.51.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
plotly==5.19.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.2
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
s3transfer==0.10.1
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.33.0
streamlit_folium==0.19.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vaderSentiment==3.3.2
wcwidth==0.2.13
xyzservices==2024.4.0
